apps/api/pyproject.toml
[project]
name = "codebase-explainer-api"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "motor>=3.5",          # MongoDB async driver
  "loguru>=0.7",
  "google-genai",
  "transformers>=4.41",
  "torch",
  "sentencepiece",
  "accelerate",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100